Don't use `grep` to filter results. Fix #3
Did a major re-write in which filtering will be done directly by
`global`. Originally, we didn't want this method because we had to
list *all* tags when inputting an empty string (and thus, have to wait
for *all* tags to be listed before proceeding, which would take
several seconds).

However, listing-all-tags-because-empty-input is solved by doing `ivy-more-chars`